src/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-jazzmin==3.0.1
django-nested-admin==4.1.6
django-timezone-field==7.1
flake8==7.3.0
gunicorn==23.0.0
idna==3.11
isort==7.0.0
kombu==5.5.4
loguru==0.7.3
mccabe==0.7.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pyTelegramBotAPI==4.29.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.1.0
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
win32_setctime==1.2.0
